fixing MPI compilation

+++ sst/elements/portals4_sm/trig_cpu/apps/ping_pong.h
@@ -74,7 +74,7 @@ public:
             ptl->PtlMEAppend(0, me, PTL_PRIORITY_LIST, NULL, me_handle);
 
 	    for ( int i = 0; i < PP_BUF_SIZE; i++ ) {
-		printf("%5d: start -> send_buffer[%d] = %llu   recv_buffer[%d] = %llu\n",my_id,i,send_buffer[i],i,recv_buffer[i]);
+		printf("%5d: start -> send_buffer[%d] = %llu   recv_buffer[%d] = %llu\n",my_id,i,(long long unsigned)send_buffer[i],i,(long long unsigned)recv_buffer[i]);
 	    }
 
 	    
@@ -101,7 +101,7 @@ public:
 	case 6:
 	    if ( ptl->PtlCTWait(ct_handle,2) ) {
 		for ( int i = 0; i < PP_BUF_SIZE; i++ ) {
-		    printf("%5d: end -> send_buffer[%d] = %llu   recv_buffer[%d] = %llu\n",my_id,i,send_buffer[i],i,recv_buffer[i]);
+		    printf("%5d: end -> send_buffer[%d] = %llu   recv_buffer[%d] = %llu\n",my_id,i,(long long unsigned)send_buffer[i],i,(long long unsigned)recv_buffer[i]);
 		}
 		uint64_t elapsed_time = cpu->getCurrentSimTimeNano()-start_time;
 		trig_cpu::addTimeToStats(elapsed_time);
