+++ sst/elements/portals4_sm/trig_cpu/apps/test_mpi.h
@@ -77,7 +77,7 @@ public:
 	case 6:
 	{
 	  for ( int i = 0; i < 32; i++ ) {
-	    printf("%5d: end -> send_buffer[%d] = %llu   recv_buffer[%d] = %llu\n",my_id,i,send_buffer[i],i,recv_buffer[i]);
+	    printf("%5d: end -> send_buffer[%d] = %llu   recv_buffer[%d] = %llu\n",my_id,i,(long long unsigned)send_buffer[i],i,(long long unsigned)recv_buffer[i]);
 	  }
 	  uint64_t elapsed_time = cpu->getCurrentSimTimeNano()-start_time;
 	  trig_cpu::addTimeToStats(elapsed_time);
